exported styled components, exported course Hero section
